removing old generator code
