+++ src/main/scala/stage/RocketChipStage.scala
@@ -2,10 +2,9 @@
 
 package freechips.rocketchip.stage
 
-import firrtl.options.{Phase, PreservesAll, Shell, Stage, StageMain}
+import firrtl.options.{Phase, PreservesAll, Shell, StageMain}
 import firrtl.stage.FirrtlCli
 import chisel3.stage.{ChiselCli, ChiselStage}
-import firrtl.AnnotationSeq
 import firrtl.options.PhaseManager.PhaseDependency
 
 class RocketChipStage extends ChiselStage with PreservesAll[Phase] {
@@ -23,20 +22,7 @@ class RocketChipStage extends ChiselStage with PreservesAll[Phase] {
     classOf[freechips.rocketchip.stage.phases.GenerateArtefacts],
   )
 
-  /* TODO: need a RunPhaseAnnotation to inject phases into ChiselStage
-  override def run(annotations: AnnotationSeq): AnnotationSeq =
-    Seq(
-      new freechips.rocketchip.stage.phases.Checks,
-      new freechips.rocketchip.stage.phases.PreElaboration,
-      new chisel3.stage.ChiselStage,
-      new freechips.rocketchip.stage.phases.GenerateFirrtl,
-      new freechips.rocketchip.stage.phases.GenerateFirrtlAnnos,
-      new freechips.rocketchip.stage.phases.GenerateTestSuiteMakefrags,
-      new freechips.rocketchip.stage.phases.GenerateROMs,
-      new freechips.rocketchip.stage.phases.GenerateArtefacts,
-    ).foldLeft(annotations){ case (a, p) => p.transform(a) }
-
-   */
+  // TODO: need a RunPhaseAnnotation to inject phases into ChiselStage
 }
 
 object RocketChipMain extends StageMain(new RocketChipStage)
