+++ src/main/scala/stage/phases/GenerateFirrtl.scala
@@ -11,6 +11,7 @@ import firrtl.AnnotationSeq
 import firrtl.options.{Phase, PreservesAll, StageOptions}
 import firrtl.options.Viewer.view
 import freechips.rocketchip.stage.RocketChipOptions
+import freechips.rocketchip.util.HasRocketChipStageUtils
 
 class GenerateFirrtl extends Phase with PreservesAll[Phase] with HasRocketChipStageUtils {
