+++ src/main/scala/stage/phases/GenerateROMs.scala
@@ -8,6 +8,7 @@ import firrtl.AnnotationSeq
 import firrtl.options.{Phase, PreservesAll, StageOptions}
 import firrtl.options.Viewer.view
 import freechips.rocketchip.stage.RocketChipOptions
+import freechips.rocketchip.util.HasRocketChipStageUtils
 
 class GenerateROMs extends Phase with PreservesAll[Phase] with HasRocketChipStageUtils {
