+++ src/main/scala/unittest/Generator.scala
@@ -2,9 +2,7 @@
 
 package freechips.rocketchip.unittest
 
-object Generator extends freechips.rocketchip.util.GeneratorApp {
-  generateFirrtl
-  generateAnno
-  generateTestSuiteMakefrags // TODO: Needed only for legacy make targets
-  generateArtefacts
-}
+import firrtl.options.StageMain
+import freechips.rocketchip.stage.RocketChipStage
+
+object Generator extends StageMain(new RocketChipStage)
